feat: Add details contact page
